move timer_device_list files out of clock
