+++ ch22-core/src/address_map/io_space/tests.rs
@@ -2,9 +2,9 @@ use std::cell::RefCell;
 use std::rc::Rc;
 
 use super::*;
-use crate::clock::timer_device_list::TimerDeviceList;
 use crate::clock::*;
 use crate::devices::io_device_mock::*;
+use crate::devices::timer_device_list::TimerDeviceList;
 use crate::interrupt_type::*;
 
 use DeviceSpeed::*;
