+++ ch22-core/src/devices/timer_device_list.rs
