+++ ch22-core/src/system.rs
@@ -4,13 +4,13 @@ use wasm_bindgen::prelude::*;
 use crate::address_map::AddressMap;
 use crate::address_map::io_space::DeviceSpeed;
 use crate::address_map::io_space::io_device_list::IODeviceID;
-use crate::clock::timer_device_list::TimerDeviceID;
-use crate::clock::timer_device_list::TimerDeviceList;
 use crate::cpu::*;
 use crate::cycle_manager::*;
 use crate::devices::js_io_device::JsIODevice;
 use crate::devices::js_timer_device::*;
 use crate::devices::static_device::StaticDevice;
+use crate::devices::timer_device_list::TimerDeviceID;
+use crate::devices::timer_device_list::TimerDeviceList;
 use crate::interrupt_type::InterruptType;
 use crate::utils;
 use crate::video::field_data::Field;
